feature/82 refactor: Exception네이밍 수정
 - Owner -> Author

+++ src/main/java/com/grasshouse/dorandoran/comment/service/CommentLikeService.java
@@ -7,6 +7,7 @@ import com.grasshouse.dorandoran.comment.repository.CommentRepository;
 import com.grasshouse.dorandoran.comment.service.dto.CommentLikeCreateRequest;
 import com.grasshouse.dorandoran.common.exception.CommentLikeAlreadyExistsException;
 import com.grasshouse.dorandoran.common.exception.CommentLikeNotFoundException;
+import com.grasshouse.dorandoran.common.exception.CommentLikerMisMatchException;
 import com.grasshouse.dorandoran.common.exception.CommentNotFoundException;
 import com.grasshouse.dorandoran.member.domain.Member;
 import lombok.RequiredArgsConstructor;
@@ -46,7 +47,9 @@ public class CommentLikeService {
     public void deleteCommentLike(Long commentLikeId, Member member) {
         CommentLike commentLike = commentLikeRepository.findById(commentLikeId)
             .orElseThrow(CommentLikeNotFoundException::new);
-
+        if (!commentLike.isSameLiker(member)) {
+            throw new CommentLikerMisMatchException();
+        }
         commentLikeRepository.delete(commentLike);
     }
 }
